pyproject.toml
[project]
name = "notion-bot"
version = "0.1.0"
authors = [
    { name = "Angel Merino", email = "acedeno00@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "datetime>=6.0",
    "dotenv>=0.9.9",
    "fastapi>=0.135.1",
    "ollama>=0.6.1",
    "pydantic>=2.12.5",
    "python-toon>=0.1.3",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "uvicorn>=0.42.0",
]

[project.scripts]
notion-bot = "notion_bot:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
